agregando screen de registros, modularizando el menuList y arreglando el merge del stash por primera vez
