+++ frontend/ballots_template_flutter/lib/routes/app_routes.dart
@@ -2,6 +2,7 @@ import 'package:ballots_template_flutter/screens/auth/login_screen.dart';
 import 'package:ballots_template_flutter/screens/invoice/invoice_screens.dart';
 import 'package:ballots_template_flutter/screens/invoice/product_invoice_screen.dart';
 import 'package:ballots_template_flutter/screens/invoice/service_invoice_screen.dart';
+import 'package:ballots_template_flutter/screens/registrations/registrations_screen.dart';
 import 'package:get/get.dart';
 import 'package:ballots_template_flutter/screens/form_register_screens.dart';
 import 'package:ballots_template_flutter/screens/home_screen.dart';
@@ -13,6 +14,7 @@ class AppRoutes {
   static const invoice = '/invoice';
   static const productInvoice = '/invoice/product';
   static const serviceInvoice = '/invoice/service';
+  static const registrations = '/registrations';
 
   static final routes = [
     GetPage(name: initial, page: () => const LoginScreen()),
@@ -21,5 +23,6 @@ class AppRoutes {
     GetPage(name: invoice, page: () => const InvoiceScreen()),
     GetPage(name: productInvoice, page: () => const ProductInvoiceScreen()),
     GetPage(name: serviceInvoice, page: () => const ServiceInvoiceScreen()),
+    GetPage(name: registrations, page: () => const RegistrationsScreen()),
   ];
 }
