+++ frontend/ballots_template_flutter/lib/screens/invoice/invoice_screens.dart
@@ -1,5 +1,6 @@
 import 'package:ballots_template_flutter/routes/app_routes.dart';
 import 'package:ballots_template_flutter/theme/colors.dart';
+import 'package:ballots_template_flutter/widgets/menu_list.dart';
 import 'package:ballots_template_flutter/widgets/screen_container.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
@@ -32,8 +33,9 @@ class InvoiceScreen extends StatelessWidget {
     return ScreenContainer(
       appBarChildren: Text(
         'Recibo',
-        style: theme.headlineLarge
-            ?.copyWith(color: AppColors.whiteColor,),
+        style: theme.headlineLarge?.copyWith(
+          color: AppColors.whiteColor,
+        ),
       ),
       children: Container(
         padding: const EdgeInsets.all(10),
@@ -43,45 +45,10 @@ class InvoiceScreen extends StatelessWidget {
               'Crear recibos de productos o servicios',
               style: theme.titleLarge,
             ),
-            SizedBox(
+            const SizedBox(height: 20),
+            Container(
               height: 200,
-              child: ListView.separated(
-                itemCount: 2,
-                itemBuilder: (context, index) {
-                  final item = list[index];
-                  String title = item['title'] ?? '';
-                  String subtitle = item['subtitle'] ?? '';
-                  return GestureDetector(
-                    onTap: item['onTap'],
-                    child: ListTile(
-                      shape: const BeveledRectangleBorder(
-                        borderRadius: BorderRadius.all(Radius.circular(10)),
-                        side: BorderSide(width: 1, color: AppColors.greyColor),
-                      ),
-                      tileColor: AppColors.cardColorSecondary,
-                      title: Text(
-                        title,
-                        style: const TextStyle(
-                          fontWeight: FontWeight.bold,
-                          fontSize: 16,
-                        ),
-                      ),
-                      subtitle: Text(
-                        subtitle,
-                        style: const TextStyle(
-                          fontWeight: FontWeight.bold,
-                          fontSize: 11,
-                        ),
-                      ),
-                      leading: Icon(item['icon']),
-                      trailing: const Icon(Icons.arrow_forward_ios),
-                    ),
-                  );
-                },
-                separatorBuilder: (context, index) {
-                  return const SizedBox(height: 12);
-                },
-              ),
+              child: MenuList(list: list),
             ),
           ],
         ),
