+++ frontend/ballots_template_flutter/lib/utils/homescreen_datalist.dart
@@ -1,6 +1,32 @@
+import 'package:ballots_template_flutter/controllers/form_controller.dart';
 import 'package:ballots_template_flutter/routes/app_routes.dart';
+import 'package:ballots_template_flutter/widgets/warning_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
+import 'package:google_fonts/google_fonts.dart';
+
+void navigateToAnotherScreen(action, String title) {
+  final FormController formController = Get.find();
+
+  if (formController.isFormValid || title == 'Ajustes') {
+    action();
+  } else {
+    Get.defaultDialog(
+      title: '',
+      titleStyle: GoogleFonts.onest(
+        fontSize: 1,
+      ),
+      content: WarningDialog(
+        title: '¡Aplicación no configurada!',
+        titleBtn: "Ir a configurar",
+        onPress: () {
+          Get.back();
+          Get.toNamed(AppRoutes.formRegister);
+        },
+      ),
+    );
+  }
+}
 
 getListTileData() {
   List<Map<String, dynamic>> listTileData = [
@@ -10,7 +36,7 @@ getListTileData() {
       'icon': Icons.receipt_long_outlined,
       "selected": true,
       "onTap": () {
-        Get.toNamed(AppRoutes.invoice);
+        navigateToAnotherScreen(() => Get.toNamed(AppRoutes.invoice), 'Recibo');
       }
     },
     {
@@ -18,63 +44,82 @@ getListTileData() {
       'subtitle': 'Crear recibo del servicio',
       'icon': Icons.receipt_rounded,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Recibo del servicio');
+      }
     },
     {
       'title': 'Recibo del producto',
       'subtitle': 'Crear recibo del producto',
       'icon': Icons.receipt_rounded,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Recibo del producto');
+      }
     },
     {
       'title': 'Registros',
       'subtitle': 'Registros de productos y servicios',
       'icon': Icons.content_paste_rounded,
       "selected": true,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(
+            () => Get.toNamed(AppRoutes.registrations), 'Registros');
+      }
     },
     {
       'title': 'Registro del producto',
       'subtitle': 'Registrar producto',
       'icon': Icons.content_paste_go,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Registro del producto');
+      }
     },
     {
       'title': 'Registro de servicio',
       'subtitle': 'Registrar servicio',
       'icon': Icons.content_paste_go,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Registro de servicio');
+      }
     },
     {
       'title': 'Lista de Productos / Servicios',
       'subtitle': 'Lista de productos y servicios',
       'icon': Icons.content_paste_search,
       "selected": true,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Lista de Productos / Servicios');
+      }
     },
     {
       'title': 'Lista de Productos',
       'subtitle': 'Listar todos los productos',
       'icon': Icons.list,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Lista de Productos');
+      }
     },
     {
       'title': 'Lista de servicios',
       'subtitle': 'Listar todos los servicios',
       'icon': Icons.list,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Lista de servicios');
+      }
     },
     {
       'title': 'Presupuestos',
       'subtitle': 'Presupuestos de productos y servicios',
       'icon': Icons.calculate_outlined,
       "selected": true,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Presupuestos');
+      }
     },
     {
       'title': 'Clientes',
@@ -88,35 +133,45 @@ getListTileData() {
       'subtitle': 'Historial de productos y servicios',
       'icon': Icons.history,
       "selected": true,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Historial');
+      }
     },
     {
       'title': 'Historial de recepción de productos',
       'subtitle': 'Historial de recepción de productos',
       'icon': Icons.history,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Historial de recepción de productos');
+      }
     },
     {
       'title': 'Historial de recibos de servicio',
       'subtitle': 'Historial de recepción de servicios',
       'icon': Icons.history,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Historial de recibos de servicio');
+      }
     },
     {
       'title': 'Historial de presupuesto',
       'subtitle': 'Presupuesto de productos o s...',
       'icon': Icons.history,
       "selected": false,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Historial de presupuesto');
+      }
     },
     {
       'title': 'Planes Premium',
       'subtitle': 'Conviértase en un usuario premium y 0...',
       'icon': Icons.workspace_premium,
       "selected": true,
-      "onTap": () {}
+      "onTap": () {
+        navigateToAnotherScreen(() {}, 'Planes Premium');
+      }
     },
     {
       'title': 'Ajustes',
@@ -124,8 +179,10 @@ getListTileData() {
       'icon': Icons.settings_sharp,
       "selected": true,
       "onTap": () {
-        // Get.to(() => const FormRegisterScreen());
-        Get.toNamed('/form-register');
+        navigateToAnotherScreen(
+          () => Get.toNamed(AppRoutes.formRegister),
+          'Ajustes',
+        );
       }
     },
   ];
